Move loadProperties methods to PropertiesUtils
