+++ java/clients/src/test/java/sleeper/clients/deploy/CdkDeployInstanceTest.java
@@ -25,7 +25,7 @@ import java.nio.file.Path;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.configuration.properties.InstancePropertiesTestHelper.createTestInstanceProperties;
-import static sleeper.configuration.properties.SleeperProperties.loadProperties;
+import static sleeper.configuration.properties.PropertiesUtils.loadProperties;
 import static sleeper.utils.RunCommandTestHelper.commandRunOn;
 
 class CdkDeployInstanceTest {
