+++ java/configuration/src/main/java/sleeper/configuration/properties/local/LoadLocalProperties.java
@@ -30,7 +30,7 @@ import java.util.Properties;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static sleeper.configuration.properties.SleeperProperties.loadProperties;
+import static sleeper.configuration.properties.PropertiesUtils.loadProperties;
 
 public class LoadLocalProperties {
