+++ java/configuration/src/test/java/sleeper/configuration/properties/format/SleeperPropertiesPrettyPrinterTest.java
@@ -41,7 +41,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.configuration.properties.SleeperProperties.loadProperties;
+import static sleeper.configuration.properties.PropertiesUtils.loadProperties;
 import static sleeper.configuration.properties.table.TablePropertiesTestHelper.createTestTableProperties;
 
 class SleeperPropertiesPrettyPrinterTest {
